lambda/requirements.txt
boto3
requests
PyJWT